add link to / and wrap profile link in suspense
